SpeechToPictogram/requirements.txt
matplotlib==3.5.2
Pillow==9.2.0
git+https://github.com/openai/whisper.git
nltk==3.7
requests==2.28.1
streamlit==1.14
ffmpeg-python
pydub==0.25.1
setuptools-rust==1.5.2
openai
gradio==3.9.1
